requirements.txt
# requirements.txt
streamlit
st-annotated-text